% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_files.R
\name{.check_files}
\alias{.check_files}
\title{Check files argument input}
\usage{
.check_files(files)
}
\arguments{
\item{files}{character}
}
\value{
character
}
\description{
Check files argument input
}
\examples{
\dontrun{.check_files(files)}
}
\keyword{internal}
