% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_files.R
\name{.check_patterns}
\alias{.check_patterns}
\title{Check patterns argument input}
\usage{
.check_patterns(patterns)
}
\arguments{
\item{patterns}{character}
}
\value{
character
}
\description{
Check patterns argument input
}
\examples{
\dontrun{.check_patterns(patterns)}
}
\keyword{internal}
