% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_files.R
\name{.files_in_directory}
\alias{.files_in_directory}
\title{List files in directory}
\usage{
.files_in_directory(dir, recursive = TRUE)
}
\arguments{
\item{files}{string}
}
\value{
character
}
\description{
List files in directory
}
\examples{
\dontrun{.files_in_directory(dir)}
}
\keyword{internal}
