% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_files.R
\name{.find_matches}
\alias{.find_matches}
\title{Search file for regular expression matches}
\usage{
.find_matches(file, pattern, encoding = "unknown", ignore.case = TRUE)
}
\arguments{
\item{file}{character}

\item{pattern}{characater}

\item{encoding}{characater}
}
\value{
data.frame
}
\description{
Search file for regular expression matches
}
\examples{
\dontrun{.find_matches(file, pattern)}
}
\keyword{internal}
