% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_files.R
\name{scan_files}
\alias{scan_files}
\title{Scan files using using regular expressions.}
\usage{
scan_files(
  files,
  patterns,
  file_type = NULL,
  encoding = "unknown",
  ignore.case = TRUE
)
}
\arguments{
\item{files}{character}

\item{patterns}{character}

\item{file_type}{character}

\item{encoding}{character}

\item{ignore.case}{boolean}
}
\value{
data.frame
}
\description{
Scan files using using regular expressions.
}
\examples{
\dontrun{scan_files("file-path", "regular-expression")}
}
